requirements.txt
gradio
google-generativeai
transformers
torch
python-dotenv
PyPDF2
latex